only users can delete their tinyURLs

+++ express_server.js
@@ -13,12 +13,16 @@ app.use(bodyParser.urlencoded({extended: true}));
 app.use(cookieParser());
 
 //DATABASES
-const userDatabase = {
+const userDatabase = { //structure of database
+  // {
+  // O1qFUflm: { id: 'O1qFUflm', email: 'lukeberzins16@gmail.com',password: 'chocolate_chip' },
+  // cwW921dh: { id: 'cwW921dh', email: 'funky-chicken-234@hotmail.com', password: 'pancakes' }
+  // }
 };
 
-const urlDatabase = {
-  b6UTxQ: { longURL: "https://www.tsn.ca", userID: "aJ48lW" },
-  i3BoGr: { longURL: "https://www.google.ca", userID: "aJ48lW" }
+const urlDatabase = { //structure of database
+  // b6UTxQ: { longURL: "https://www.clubpenguin.ca", userID: "aJ48lW" },
+  // i3BoGr: { longURL: "https://www.google.ca", userID: "aJ48lW" }
 };
 
 //FUNCTIONS
@@ -29,20 +33,20 @@ const generateRandomString = (name) => {
   while (result.length < name) {
     result += string[Math.floor((Math.random() * 61))];
   }
-  return result; //not the most elegant but I came up with it myself, so come on, you know?
+  return result; //not the most elegant but who is?
 };
 
-//URL DATABASE FUNCTIONS
 const deleteItem = (database, key) => {
   delete database[key];
 };
 
-const editItem = (database, key, long, userInfo) => {
-  console.log(userInfo)
-  database[key] = {longURL: long, userID: userInfo }
+//URL DATABASE FUNCTIONS
+
+const editItem = (database, key, long, userInfo) => { //specific to URL database structure
+  database[key] = {longURL: long, userID: userInfo };
 };
 
-const urlsForUser = id => {
+const urlsForUser = id => { //returns an object with the urlDatabase key-values that match the specified user id
   let songTags = {};
   for (let song in urlDatabase) {
     if (urlDatabase[song]["userID"] === id) {
@@ -50,7 +54,7 @@ const urlsForUser = id => {
     }
   }
   return songTags;
-} 
+};
 
 //  USER DATABASE FUNCTIONS
 const createUser = (name, pass) => {
@@ -60,18 +64,21 @@ const createUser = (name, pass) => {
     email: name,
     password: pass
   };
+  console.log(userDatabase);
   return created; //return the newly created user to use in automatic login after registration
 };
 
-const checkUser = (field, newUser) => {
+const checkUser = (field, newUser) => { //this checks user data against register and login queries
   let value;
   for (let userKnown in userDatabase) {
-    value = (userDatabase[userKnown]) 
+    value = (userDatabase[userKnown]);
     if (userDatabase[userKnown][field] === newUser) {
-      return value; //If there's a key-value that matches the searched one, (for /register the searched one is the 
-    }              // user's requested email), then return true
+      return value; //if they do exist return their info for /login
+      //if they do exist give a truthy value for /register
+    }
   }
-  return false; //if newUser value doesnt exist in userDatabase then return false
+  //if newUser value doesnt exist in userDatabase then return false for /login
+  return false; //if newUser value doesnt exist in userDatabase then return false for /register
 };
 
 // APP GETS
@@ -92,10 +99,9 @@ app.get("/register", (req, res) => {
 
 app.get("/urls", (req, res) => {
   if (req.cookies["user_id"]) {
-  const urlList = urlsForUser(req.cookies["user_id"]["id"]) 
-  console.log(urlList)
-  const templateVars = { user: req.cookies["user_id"], urls: urlList };
-  res.render("urls_index", templateVars);
+    const urlList = urlsForUser(req.cookies["user_id"]["id"]);
+    const templateVars = { user: req.cookies["user_id"], urls: urlList };
+    res.render("urls_index", templateVars);
   } else {
     res.redirect("/login");
   }
@@ -108,15 +114,14 @@ app.get("/urls", (req, res) => {
 
 app.get("/urls/new", (req, res) => {
   if (req.cookies["user_id"]) {
-  const templateVars = { user: req.cookies["user_id"]};
-  res.render("urls_new", templateVars);
+    const templateVars = { user: req.cookies["user_id"]};
+    res.render("urls_new", templateVars);
   } else {
     res.redirect("/login");
-  };
+  }
 });
 
 app.get("/u/:shortURL", (req, res) => {
-  console.log(req.params.shortURL, "/u/short")
   const longURL = urlDatabase[req.params.user_id]["longURL"];
   res.redirect(longURL);
 });
@@ -184,7 +189,9 @@ app.post("/urls", (req, res) => {
 });
 
 app.post("/urls/:shortURL/delete", (req, res) => {
-  deleteItem(urlDatabase, req.params.shortURL);
+  if (req.cookies["user_id"]["id"] === urlDatabase[req.params.shortURL]["userID"]) {
+    deleteItem(urlDatabase, req.params.shortURL);
+  }
   res.redirect(`/urls`);
 });
